requirements.txt
django==1.10.3
pytest-django==3.0.0
djangorestframework==3.5.3
django-debug-toolbar==1.6
flake8==3.2.0